feature/dev-cleanup: Code Structure changes, and licitaciones endpoint working READY TO DEV
