--- src/main/java/com/example/reto_backend_febrero2026/familia/mapper/FamiliaMapper.java
@@ -1,20 +1,0 @@
-package com.example.reto_backend_febrero2026.familia.mapper;
-
-import com.example.reto_backend_febrero2026.familia.Familia;
-import com.example.reto_backend_febrero2026.familia.dto.FamiliaDTO;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-
-@Mapper(componentModel = "spring")
-public interface FamiliaMapper {
-
-    @Mapping(source = "cod", target = "cod")
-    @Mapping(source = "descripcion", target = "descripcion")
-    FamiliaDTO familyToFamilyDTO(Familia family);
-
-    @Mapping(source = "cod", target = "cod")
-    @Mapping(source = "descripcion", target = "descripcion")
-    Familia familyDTOtoFamily (FamiliaDTO familyDTO);
-
-}
