+++ src/main/java/com/example/reto_backend_febrero2026/integration/servlet/controller/TestController.java
@@ -1,8 +1,7 @@
 package com.example.reto_backend_febrero2026.integration.servlet.controller;
 
-import com.example.reto_backend_febrero2026.integration.servlet.dto.LicitacionItemRecord;
 import com.example.reto_backend_febrero2026.integration.servlet.service.ArceClientService;
-import com.example.reto_backend_febrero2026.licitacion.dto.LicitacionDTO;
+import com.example.reto_backend_febrero2026.licitacion.LicitacionDTO;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
