+++ src/main/java/com/example/reto_backend_febrero2026/licitacion/ILicitacionRepository.java
@@ -1,0 +1,7 @@
+package com.example.reto_backend_febrero2026.licitacion;
+
+public interface ILicitacionRepository {
+    Licitacion getLicitacionById(int licitacionId);
+    Licitacion save(Licitacion licitacion);
+    Licitacion getLicitacionByTitulo(String titulo);
+}
